Don't rely on platform encoding being UTF-8
The input stream reader that reads responses from the Aylien API
doesn’t specify a character encoding. This means it uses platform
encoding (specified by -Dfile.encoding=XXXX).

Everything works fine when the platform encoding is UTF-8, but when it
is not the UTF-8 responses from the API are incorrectly interpreted.
This fixes.